Update sys-dm-pdw-exec-sessions-transact-sql.md
When you create a private endpoint, the traffic flows through the azure backend network. When the packet reaches the SQL db, it emanates from a network device that applies IpV6 ip address on the packet. This IpV6 address range is an internal range and no public IpV6 address will have this value. That is how SQL knows that the traffic is coming from a private endpoint.